app/requirements.txt
flask
pyjwt
requests
cryptography
prometheus_client